+++ src/neptune/integrations/mosaicml/__init__.py
@@ -19,4 +19,4 @@ from neptune.internal.utils.requirement_check import require_installed
 
 require_installed("mosaicml")
 
-from composer.loggers import NeptuneLogger  # type: ignore[import]  # noqa: F401,F403,E402
+from composer.loggers import NeptuneLogger  # noqa: F401,F403,E402